requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
websockets==12.0
aiofiles==23.2.1
gunicorn==21.2.0
Pillow==10.2.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
structlog==24.1.0
cryptography==42.0.0
pyyaml==6.0.1
kubernetes==29.0.0
resend==0.8.0
